Add isValid for the schema being used in the UI rather than the previous schema used for the name note
