+++ src/database/DataTypes/NameNote.js
@@ -1,14 +1,5 @@
 import Realm from 'realm';
-import Ajv from 'ajv';
 
-const ajvOptions = {
-  errorDataPath: 'property',
-  allErrors: true,
-  multipleOfPrecision: 8,
-  schemaId: 'auto',
-  unknownFormats: 'ignore',
-  jsonPointers: true,
-};
 export class NameNote extends Realm.Object {
   get data() {
     try {
@@ -24,17 +15,6 @@ export class NameNote extends Realm.Object {
     this._data = JSON.stringify(newValue);
   }
 
-  get isValid() {
-    const { jsonSchema } = this.formSchema ?? {};
-
-    if (!jsonSchema) return false;
-
-    const ajv = new Ajv(ajvOptions);
-    const result = ajv.validate(jsonSchema, this.data);
-
-    return result;
-  }
-
   toObject() {
     return {
       id: this.id,
@@ -42,7 +22,6 @@ export class NameNote extends Realm.Object {
       data: this.data,
       nameID: this.name.id,
       patientEventID: this.patientEvent.id,
-      isValid: this.isValid,
     };
   }
 }
@@ -56,7 +35,6 @@ NameNote.schema = {
     _data: { type: 'string', optional: true },
     name: 'Name',
     patientEvent: 'PatientEvent',
-    formSchema: { type: 'FormSchema', optional: true },
   },
 };
